+++ src/main/java/keeper/project/homepage/admin/controller/library/BookManageController.java
@@ -1,7 +1,5 @@
-package keeper.project.homepage.controller.admin.library;
+package keeper.project.homepage.admin.controller.library;
 
-import java.util.List;
-import javax.servlet.ServletRequest;
 import javax.servlet.http.HttpServletRequest;
 import keeper.project.homepage.common.ImageCenterCrop;
 import keeper.project.homepage.dto.library.BookDto;
@@ -10,12 +8,10 @@ import keeper.project.homepage.entity.FileEntity;
 import keeper.project.homepage.entity.ThumbnailEntity;
 import keeper.project.homepage.dto.result.ListResult;
 import keeper.project.homepage.entity.library.BookBorrowEntity;
-import keeper.project.homepage.entity.library.BookEntity;
 import keeper.project.homepage.exception.CustomAboutFailedException;
-import keeper.project.homepage.repository.library.BookRepository;
 import keeper.project.homepage.service.FileService;
 import keeper.project.homepage.service.ThumbnailService;
-import keeper.project.homepage.service.library.BookManageService;
+import keeper.project.homepage.admin.service.library.BookManageService;
 import keeper.project.homepage.service.ResponseService;
 import keeper.project.homepage.service.util.AuthService;
 import lombok.RequiredArgsConstructor;
@@ -23,17 +19,12 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
