+++ src/main/java/keeper/project/homepage/user/controller/library/LibraryMainController.java
@@ -1,19 +1,15 @@
-package keeper.project.homepage.controller.library;
+package keeper.project.homepage.user.controller.library;
 
-import java.util.List;
 import keeper.project.homepage.dto.result.ListResult;
 import keeper.project.homepage.dto.result.SingleResult;
 import keeper.project.homepage.entity.library.BookEntity;
 import keeper.project.homepage.service.ResponseService;
-import keeper.project.homepage.service.library.LibraryMainService;
+import keeper.project.homepage.user.service.library.LibraryMainService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.HttpStatus;
-import org.springframework.data.domain.Page;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
