+++ src/main/java/keeper/project/homepage/user/service/library/LibraryMainService.java
@@ -1,4 +1,4 @@
-package keeper.project.homepage.service.library;
+package keeper.project.homepage.user.service.library;
 
 import java.util.ArrayList;
 import java.util.HashSet;
